+++ db.json
@@ -40,7 +40,7 @@
       "currency": "XLM",
       "hold": "0",
       "pending_balance": "0",
-      "balance": "0.2",
+      "balance": "0",
       "name": "Stellar",
       "type": "digital",
       "imgURL": "https://res.cloudinary.com/dwoc5fknz/image/upload/v1593000381/alice_v3/XLM.svg",
