+++ src/features/wallets/components/AddNewWallet.tsx
@@ -7,6 +7,7 @@ import Loader from "../../../components/shared/Loader";
 import { ErrorState } from "../../../components/shared/ErrorState";
 import { useState } from "react";
 import Alert from "../../../components/shared/Alert";
+import { useAddWallet } from "../api/add-wallet";
 
 interface Props {
   isOpen: boolean;
@@ -17,50 +18,13 @@ interface Props {
 
 export function AddNewWallet({ isOpen, title, description, onClose }: Props) {
   const { wallets, loading, error, fetchWallets } = useGetWallets();
-
-  const [selectedWallet, setSelectedWallet] = useState<string>("");
-  const [alertState, setAlertState] = useState<"success" | "error" | null>(null);
-  const [alertMessage, setAlertMessage] = useState<string>("");
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-
+  const { isLoading, alertMessage, alertState, setSelectedWallet, handleSubmit } = useAddWallet({ onClose });
   // Handle select change
   const handleSelectChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
     setSelectedWallet(event.target.value);
   };
   const BASE_URL = 'http://localhost:3090';
-  // Handle form submission
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setIsLoading(true);
-    try {
-      const response = await fetch(`${BASE_URL}/accounts`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ wallet: selectedWallet }),
-      });
-  
-      if (!response.ok) {
-        setAlertState('error');
-        setAlertMessage('Network error');
-      } else {
-        // Handle successful creation
-        setAlertState('success');
-        setAlertMessage('Wallet created successfully!');
-        // Close the modal after a successful creation
-        onClose(); // Close the modal
-      }
-    } catch (error: any) {
-      console.error(error);
-      setAlertState('error');
-      setAlertMessage('Error creating wallet: ' + error.message);
-    } finally {
-      setIsLoading(false); // Ensure this is in a finally block to always run
-    }
-  };
-  
-  
+  // Handle form submission  
 
   // Map the wallets to the expected format for the SelectField options
   const walletOptions = wallets.map(wallet => ({
